refactor: update import paths for technical indicators to use backend references
